requirements.txt
feedparser==5.1.3
pushbullet.py==0.5.0
configparser==3.5.0b2
